Move PromisesResolver to ES6
